requirements.txt
aiohttp
asyncpg
SQLAlchemy==1.4.49
python-dotenv
bcrypt
